各ページで発生していた、nullに対するaddEventListener使用によるエラーを解消
