+++ app/javascript/events.js
@@ -9,23 +9,4 @@ document.addEventListener('DOMContentLoaded', () => {
       render: (h) => h(Events)
     }).$mount(selector)
   }
-
-  document
-    .getElementById('event_start_at')
-    .addEventListener('blur', (event) => {
-      const eventStartAtDate = event.target.value
-      const idsToSubstituteDate = ['event_end_at', 'event_open_end_at']
-      idsToSubstituteDate.forEach((idToSubstituteDate) => {
-        substituteDate(
-          eventStartAtDate,
-          document.getElementById(idToSubstituteDate)
-        )
-      })
-    })
-
-  function substituteDate(date, destElement) {
-    if (destElement.value === '') {
-      destElement.value = date
-    }
-  }
 })
