+++ app/javascript/packs/application.js
@@ -34,6 +34,7 @@ import '../learning'
 import '../learning-status'
 import '../markdown'
 import '../mentor-mode'
+import '../new-event-date-set'
 import '../niconico_calendar'
 import '../no_learn'
 import '../notifications'
